[FIX] 도감 수집률 및 무드타입 해금 로직 수정

+++ src/main/java/com/example/moodtail/domain/cocktail/service/CocktailService.java
@@ -142,7 +142,7 @@ public class CocktailService {
         int totalCocktailCount = cocktails.size();
         int unlockedCocktailCount = unlockedCocktailIds.size();
 
-        int collectionRate = moodTypeCalculator.calculateCollectionRate(
+        double collectionRate = moodTypeCalculator.calculateCollectionRate(
                 unlockedCocktailCount,
                 totalCocktailCount
         );
